Fix Prometheus Exporter TLS. Add tests
