+++ controllers/solrcloud_controller_externaldns_test.go
@@ -102,7 +102,7 @@ var _ = FDescribe("SolrCloud controller - External DNS", func() {
 			// Env Variable Tests
 			expectedEnvVars := map[string]string{
 				"ZK_HOST":        "host:7271/",
-				"SOLR_HOST":      "$(POD_HOSTNAME)." + solrCloud.Namespace + "." + testDomain,
+				"SOLR_HOST":      "$(POD_NAME)." + solrCloud.Namespace + "." + testDomain,
 				"SOLR_PORT":      "3000",
 				"SOLR_NODE_PORT": "3000",
 				"SOLR_OPTS":      "-DhostPort=$(SOLR_NODE_PORT)",
@@ -182,7 +182,7 @@ var _ = FDescribe("SolrCloud controller - External DNS", func() {
 			// Env Variable Tests
 			expectedEnvVars := map[string]string{
 				"ZK_HOST":        "host:7271/",
-				"SOLR_HOST":      "$(POD_HOSTNAME)." + solrCloud.HeadlessServiceName() + "." + solrCloud.Namespace,
+				"SOLR_HOST":      "$(POD_NAME)." + solrCloud.HeadlessServiceName() + "." + solrCloud.Namespace,
 				"SOLR_PORT":      "2000",
 				"SOLR_NODE_PORT": "2000",
 				"SOLR_OPTS":      "-DhostPort=$(SOLR_NODE_PORT)",
@@ -253,7 +253,7 @@ var _ = FDescribe("SolrCloud controller - External DNS", func() {
 			// Env Variable Tests
 			expectedEnvVars := map[string]string{
 				"ZK_HOST":        "host:7271/",
-				"SOLR_HOST":      "$(POD_HOSTNAME)." + solrCloud.Namespace + "." + testDomain,
+				"SOLR_HOST":      "$(POD_NAME)." + solrCloud.Namespace + "." + testDomain,
 				"SOLR_PORT":      "3000",
 				"SOLR_NODE_PORT": "3000",
 				"SOLR_OPTS":      "-DhostPort=$(SOLR_NODE_PORT)",
@@ -323,7 +323,7 @@ var _ = FDescribe("SolrCloud controller - External DNS", func() {
 			// Env Variable Tests
 			expectedEnvVars := map[string]string{
 				"ZK_HOST":        "host:7271/",
-				"SOLR_HOST":      "$(POD_HOSTNAME)." + solrCloud.HeadlessServiceName() + "." + solrCloud.Namespace,
+				"SOLR_HOST":      "$(POD_NAME)." + solrCloud.HeadlessServiceName() + "." + solrCloud.Namespace,
 				"SOLR_PORT":      "3000",
 				"SOLR_NODE_PORT": "3000",
 				"SOLR_OPTS":      "-DhostPort=$(SOLR_NODE_PORT)",
@@ -391,7 +391,7 @@ var _ = FDescribe("SolrCloud controller - External DNS", func() {
 			// Env Variable Tests
 			expectedEnvVars := map[string]string{
 				"ZK_HOST":        "host:7271/",
-				"SOLR_HOST":      "$(POD_HOSTNAME)." + solrCloud.Namespace + "." + testDomain,
+				"SOLR_HOST":      "$(POD_NAME)." + solrCloud.Namespace + "." + testDomain,
 				"SOLR_PORT":      "3000",
 				"SOLR_NODE_PORT": "3000",
 				"SOLR_OPTS":      "-DhostPort=$(SOLR_NODE_PORT)",
@@ -471,7 +471,7 @@ var _ = FDescribe("SolrCloud controller - External DNS", func() {
 			// Env Variable Tests
 			expectedEnvVars := map[string]string{
 				"ZK_HOST":        "host:7271/",
-				"SOLR_HOST":      "$(POD_HOSTNAME)." + solrCloud.HeadlessServiceName() + "." + solrCloud.Namespace + ".svc." + testKubeDomain,
+				"SOLR_HOST":      "$(POD_NAME)." + solrCloud.HeadlessServiceName() + "." + solrCloud.Namespace + ".svc." + testKubeDomain,
 				"SOLR_PORT":      "2000",
 				"SOLR_NODE_PORT": "2000",
 				"SOLR_OPTS":      "-DhostPort=$(SOLR_NODE_PORT)",
@@ -536,7 +536,7 @@ var _ = FDescribe("SolrCloud controller - External DNS", func() {
 			// Env Variable Tests
 			expectedEnvVars := map[string]string{
 				"ZK_HOST":        "host:7271/",
-				"SOLR_HOST":      "$(POD_HOSTNAME)." + solrCloud.Namespace + "." + testDomain,
+				"SOLR_HOST":      "$(POD_NAME)." + solrCloud.Namespace + "." + testDomain,
 				"SOLR_PORT":      "2000",
 				"SOLR_NODE_PORT": "2000",
 				"SOLR_OPTS":      "-DhostPort=$(SOLR_NODE_PORT)",
