+++ tests/e2e/test_utils_test.go
@@ -455,7 +455,7 @@ func (r *ExecError) Error() string {
 }
 
 func callSolrApiInPod(ctx context.Context, solrCloud *solrv1beta1.SolrCloud, httpMethod string, apiPath string, queryParams map[string]string, hostnameOptional ...string) (response string, err error) {
-	hostname := "${POD_HOSTNAME}"
+	hostname := "${POD_NAME}"
 	if len(hostnameOptional) > 0 {
 		hostname = hostnameOptional[0]
 	}
